Added cbot_navigation package
