+++ cbot_navigation/config/params.yaml
@@ -1,0 +1,2 @@
+reference_latitude: 15.455296
+reference_longitude: 73.8025609
